Trigger/requirements.txt
asgiref==3.4.0
pika>=1.2.0
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary>=2.9
